# Reason for This PR

#2635

## Description of Changes

### Important note

In the current RFC we try to batch requests and then send only 1 notification to the guest if needed. This is not fully compliant with the virtio standard which states that we should send a notification after processing each descriptor if `queue.needs_notification()` returns true: https://docs.oasis-open.org/virtio/virtio/v1.1/csprd01/virtio-v1.1-csprd01.html#x1-400007 . But the current approach seems to work.

### Performance Results

**Test setup:** We run on a m5d.metal host with Ubuntu and kernel 5.5.19. We start a number of microVMs with 8vCPUs and 16GB of RAM running linux kernel v5.5.19 as a guest OS. Each microVMs is pinned to NUMA node 0. We mount a ramdisk. For each microVM we create an ext4 file on the ramdisk and we attach it to the microVM as /dev/vdb. We want to have the content of the drives in the page cache in order to asses the performance improvement brought by notification suppression. Otherwise we would be limited by the disk access speed. Inside the VM we run the following command:

```
fio --name=test --ioengine=libaio --rw=readwrite --randrepeat=1 --bs=4k --size=2G --numjobs=1 --iodepth=32 --runtime=30 --time_base=1 --end_fsync=1 --direct=1 --filename=/dev/vdb
```

Current implementation:
|description|IOPS|VMM thread CPU load|kworker threads CPU load|IOPS per CPU load|
|-|-|-|-|-|
|1 VMs|98.96|0.37|0.12|203.43||
|4 VMs|372.89|1.44|0.45|197.69||
|8 VMs|773.69|3.1|0.87|194.8||
|16 VMs|1400.15|6.6|2.03|162.33||
|32 VMs|2189.28|8.38|2.08|209.14||
|48 VMs|1377.53|7.0|1.96|153.83||

Notification suppression:
|description|IOPS|VMM thread CPU load|kworker threads CPU load|IOPS per CPU load|
|-|-|-|-|-|
|1 VMs|147.67|0.4|0.07|313.05||
|4 VMs|580.98|1.62|0.27|307.34||
|8 VMs|1163.95|3.42|0.55|292.82||
|16 VMs|2092.06|7.48|1.18|241.71||
|32 VMs|3587.84|10.45|1.13|309.81||
|48 VMs|1772.47|7.96|1.78|182.01||

If we look at the `IOPS per CPU load` metric, we observe an improvement of about 50% for the most cases when using notification suppression. There's a notable exception in the last case (48 VMs) where the performance improvement is only about 
20%. It would be worth investigating why this happens. If so, we'll do it in a different issue.

- [ ] This functionality can be added in [`rust-vmm`](https://github.com/rust-vmm).

## License Acceptance

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.

## PR Checklist

`[Author TODO: Meet these criteria.]`
`[Reviewer TODO: Verify that these criteria are met. Request changes if not]`

- [x] All commits in this PR are signed (`git commit -s`).
- [x] The reason for this PR is clearly provided (issue no. or explanation).
- [x] The description of changes is clear and encompassing.
- [x] Any required documentation changes (code and docs) are included in this PR.
- [x] Any newly added `unsafe` code is properly documented.
- [x] Any API changes are reflected in `firecracker/swagger.yaml`.
- [x] Any user-facing changes are mentioned in `CHANGELOG.md`.
- [x] All added/changed functionality is tested.