@serban300 2021-07-13
Closing in favour of #2658 . I removed batching for the moment in order to be fully virtio compliant.